Renames ConstraintSolver to OrToolsSolver.
