+++ tests/Dyna.Core.Tests.Unit/Dyna.Core.Tests.Unit.csproj
@@ -71,8 +71,8 @@
     <Compile Include="Models\ModelModelTests.cs" />
     <Compile Include="Models\VariableModelTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Solver\ConstraintSolverSimpleTests.cs" />
-    <Compile Include="Solver\ConstraintSolverTests.cs" />
+    <Compile Include="Solver\OrToolsSolverSimpleTests.cs" />
+    <Compile Include="Solver\OrToolsSolverTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Dyna.Core\Dyna.Core.csproj">
